refactor: extract planner to _internal module
