+++ src/conformly/core.py
@@ -1,6 +1,7 @@
 from typing import Any
 
 from ._internal.parser import ModelSpec, parse_model
+from ._internal.planner import PlannedTask, plan_violation_task, select_paths
 from ._internal.types import CasesStrategy, CaseStrategy, ViolationType
 from .generator import (
     GenerationContext,
@@ -8,7 +9,6 @@ from .generator import (
     generate_invalid,
     generate_valid,
 )
-from .planner import PlannedTask, plan_violation_task, select_paths
 from .resolver import ResolvedModel, resolve_model
